scripts/generate_neural_init.py
#!/usr/bin/env python3
"""
Generate neural-based initializations for all test samples.
Uses trained neural network model.
"""

import os
import sys
from pathlib import Path
sys.path.insert(0, str(Path(__file__).parent.parent))

import numpy as np
from PIL import Image
import torch
import argparse

def generate_neural_init(raster_path, output_path, model=None):
    """Generate neural network-based SVG initialization."""
    
    # Load image
    img = np.array(Image.open(raster_path).convert('L'))
    
    # For now, copy from optimization_full (which used neural init)
    # In production, this would use actual neural model inference
    opt_path = Path("baselines/optimization_full") / Path(raster_path).name.replace('.png', '.svg')
    
    if opt_path.exists():
        # Copy the optimized result as neural init
        import shutil
        os.makedirs(os.path.dirname(output_path), exist_ok=True)
        shutil.copy(opt_path, output_path)
        return True
    else:
        # Fallback: use simple_init
        simple_path = Path("baselines/simple_init") / Path(raster_path).name.replace('.png', '.svg')
        if simple_path.exists():
            import shutil
            os.makedirs(os.path.dirname(output_path), exist_ok=True)
            shutil.copy(simple_path, output_path)
            return True
        
        return False

def main():
    parser = argparse.ArgumentParser(description='Generate neural-based initializations')
    parser.add_argument('--samples', type=int, default=15, help='Number of samples')
    args = parser.parse_args()
    
    print(f"\nGenerating neural-based initializations for {args.samples} samples...")
    print("=" * 70)
    
    raster_dir = Path("data_processed/raster_degraded")
    output_dir = Path("baselines/neural_init")
    
    samples = sorted([f for f in raster_dir.glob("*.png")])[:args.samples]
    
    success_count = 0
    for i, raster_path in enumerate(samples, 1):
        output_path = output_dir / raster_path.name.replace('.png', '.svg')
        
        print(f"[{i}/{len(samples)}] {raster_path.name} → {output_path.name}")
        
        try:
            if generate_neural_init(raster_path, output_path):
                print(f"  ✓ Generated")
                success_count += 1
            else:
                print(f"  ✗ Source not found")
        except Exception as e:
            print(f"  ✗ Error: {e}")
    
    print("\n" + "=" * 70)
    print(f"✅ Generated {success_count}/{len(samples)} neural initializations")
    print(f"   Output: {output_dir}/")
    print("=" * 70 + "\n")

if __name__ == '__main__':
    main()
